requirements.txt
asgiref==3.8.1
blinker==1.8.2
Bootstrap-Flask==2.4.1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Flask==3.0.3
Flask-WTF==1.2.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.2
WTForms==3.2.1
gunicorn